+++ src/main/java/step2/README.md
@@ -79,4 +79,5 @@
 - [x] 당첨번호 저장 객체 생성
 - [x] 구입한 로또 번호와 당첨 번호 비교하여 비교결과 Map에 담아두기
 - [x] Map에 담긴 결과에 따라 당첨상금 계산
+- [x] View 생성
