+++ src/test/java/step2/LottoTicketTest.java
@@ -2,6 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoTicket;
 
 import java.util.Arrays;
 import java.util.List;
