+++ src/test/java/step2/LottoTicketsTest.java
@@ -2,6 +2,8 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoTicket;
+import step2.domain.LottoTickets;
 
 import java.util.Arrays;
 import java.util.List;
