+++ src/test/java/step2/PrizeMoneyTest.java
@@ -2,6 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.PrizeMoney;
 
 import java.util.HashMap;
 import java.util.Map;
