+++ src/test/java/step2/WinningLottoNumbersTest.java
@@ -2,6 +2,9 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoTickets;
+import step2.domain.PrizeMoney;
+import step2.domain.WinningLottoNumbers;
 
 import java.util.Arrays;
 import java.util.List;
